Reorganise files Move the subject set picker to @shared/components/SubjectSetPicker.
