+++ packages/app-project/src/shared/components/SubjectSetPicker/components/SubjectSetCard/SubjectSetCard.spec.js
