+++ packages/app-project/src/shared/components/WorkflowSelector/components/index.js
@@ -1,2 +1,1 @@
-export { default as SubjectSetPicker } from './SubjectSetPicker'
 export { default as WorkflowSelectButton } from './WorkflowSelectButton'
